+++ client/broadcast_test.go
@@ -7,27 +7,11 @@ import (
 
 	"github.com/cometbft/cometbft/crypto/tmhash"
 	"github.com/cometbft/cometbft/mempool"
-	"github.com/cometbft/cometbft/rpc/client/mock"
-	coretypes "github.com/cometbft/cometbft/rpc/core/types"
-	cmttypes "github.com/cometbft/cometbft/types"
 	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
-
-type MockClient struct {
-	mock.Client
-	err error
-}
-
-func (c MockClient) BroadcastTxAsync(_ context.Context, _ cmttypes.Tx) (*coretypes.ResultBroadcastTx, error) {
-	return nil, c.err
-}
-
-func (c MockClient) BroadcastTxSync(_ context.Context, _ cmttypes.Tx) (*coretypes.ResultBroadcastTx, error) {
-	return nil, c.err
-}
 
 func CreateContextWithErrorAndMode(err error, mode string) Context {
 	return Context{
@@ -61,5 +45,22 @@ func TestBroadcastError(t *testing.T) {
 			require.NotEmpty(t, resp.Codespace)
 			require.Equal(t, txHash, resp.TxHash)
 		}
+	}
+}
+
+func TestBroadcastCancellation(t *testing.T) {
+	modes := []string{
+		flags.BroadcastAsync,
+		flags.BroadcastSync,
+	}
+
+	txBytes := []byte{0xA, 0xB}
+	cmdCtx, cancel := context.WithCancel(context.Background())
+	cancel()
+
+	for _, mode := range modes {
+		ctx := CreateContextWithErrorAndMode(nil, mode).WithCmdContext(cmdCtx)
+		_, err := ctx.BroadcastTx(txBytes)
+		require.ErrorIs(t, err, context.Canceled)
 	}
 }